[224] - Fix animation duration

+++ Source/AdmiralSwiftUI/Classes/Components/Controls/SegmentControl/LogoTab/LogoTab.swift
@@ -64,7 +64,7 @@ public struct LogoTab: View {
             .foregroundColor(Color.clear)
             .frame(width: segmentSize.width, height: Constants.height)
             .offset(x: self.computeActiveSegmentHorizontalOffset(), y: 0)
-            .animation(Animation.linear(duration: Durations.Default.double))
+            .animation(Animation.linear(duration: Durations.Default.half))
             .eraseToAnyView()
         
     }
